[SEYE-4] Add services
