+++ orderflow_backend/src/main/java/com/orderflow/models/Admin.java
@@ -1,5 +1,10 @@
 package com.orderflow.models;
 
+import jakarta.persistence.Entity;
+import jakarta.persistence.Table;
+
+@Entity
+@Table
 public class Admin extends User {
 
     public Admin() {
