--- src/main/java/com/orderflow/models/Role.java
@@ -1,7 +1,0 @@
-package com.orderflow.models;
-
-public enum Role {
-    CLIENT,
-    COURIER,
-    ADMIN
-}
