--- src/main/java/com/orderflow/repositories/UserRepository.java
@@ -1,11 +1,0 @@
-package com.orderflow.repositories;
-
-import com.orderflow.models.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
-    boolean existsByEmail(String email);
-    boolean existsByUsername(String username);
-}
