--- src/main/resources/application.properties
@@ -1,18 +1,0 @@
-# ????????? ???? ?????? H2 (? ??????)
-spring.datasource.url=jdbc:h2:file:/Users/mahmudovseravgan/test;DB_CLOSE_DELAY=-1;AUTO_SERVER=TRUE
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=
-
-# ????? SQL-???????? ? ???????
-spring.jpa.show-sql=true
-
-# ?????????????? ?????????? ????? ???? ??????
-spring.jpa.hibernate.ddl-auto=update
-
-# ???????? ???????? ??? H2
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
-
-# ????????? H2 Console
-spring.h2.console.enabled=true
-spring.h2.console.path=/h2-console
