--- src/test/java/com/orderflow/ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.orderflow;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
